split Candidate struct
