+++ libakaza/src/engine/bigram_word_viterbi_engine.rs
@@ -14,8 +14,9 @@ use crate::dict::loader::load_dicts;
 use crate::dict::merge_dict::merge_dict;
 use crate::dict::skk::read::read_skkdict;
 use crate::engine::base::HenkanEngine;
+use crate::graph::candidate::Candidate;
 use crate::graph::graph_builder::GraphBuilder;
-use crate::graph::graph_resolver::{Candidate, GraphResolver};
+use crate::graph::graph_resolver::GraphResolver;
 use crate::graph::lattice_graph::LatticeGraph;
 use crate::graph::segmenter::Segmenter;
 use crate::kana_trie::cedarwood_kana_trie::CedarwoodKanaTrie;
