+++ libakaza/src/graph/mod.rs
@@ -1,3 +1,4 @@
+pub mod candidate;
 pub mod graph_builder;
 pub mod graph_resolver;
 pub mod lattice_graph;
